Core executor integration and tests
Core integration (prepared statement caching in executor):

1. Query template preservation (citus_clauses.c/h):
   - Add EVALUATE_FUNCTIONS mode to CoordinatorEvaluationMode enum
   - New ExecuteCoordinatorEvaluableFunctions() evaluates functions
     while preserving Param nodes for parameterized queries

2. Task metadata (multi_physical_planner.h, citus_custom_scan.c):
   - Add preparedStatementPlanId and jobQueryForPrepare fields to Task
   - CitusBeginReadOnlyScan and CitusBeginModifyScan save pre-evaluation
     job query and annotate tasks with plan ID for cache lookup
   - INSERT commands excluded (query trees have special RTEs that
     pg_get_query_def cannot deparse)

3. SendNextQuery prepared statement path (adaptive_executor.c):
   - New if-block at top of SendNextQuery() checks
     EnablePreparedStatementCaching && task->jobQueryForPrepare
   - Cache miss: UpdateRelationToShardNames + pg_get_query_def to
     construct parameterized shard query, PQprepare via SendRemotePrepare,
     cache the entry, then PQsendQueryPrepared via SendRemotePreparedQuery
   - Cache hit: reuse cached statement name with PQsendQueryPrepared
   - Fallback: goto plain_sql when cache is full
   - Skip MarkUnreferencedExternParams when caching enabled to preserve
     original parameter types

4. libpq wrappers (remote_commands.c/h):
   - SendRemotePrepare: synchronous PQprepare wrapper
   - SendRemotePreparedQuery: async PQsendQueryPrepared wrapper

5. Cache key fix (prepared_statement_cache.c/h):
   - Changed planId from uint32 to uint64 to match DistributedPlan.planId

Tests:

- Regression test suite for prepared statement caching (GUC toggle,
  single-shard SELECT, multi-shard routing, INSERT/UPDATE/DELETE,
  function evaluation, concurrent statements, connection loss re-prepare)

+++ src/include/distributed/prepared_statement_cache.h
@@ -14,7 +14,6 @@
 
 #include "postgres.h"
 
-#include "nodes/pg_list.h"
 #include "utils/hsearch.h"
 
 /* compile-time limit for per-connection cached prepared statements */
@@ -31,7 +30,7 @@
  */
 typedef struct PreparedStatementCacheKey
 {
-	uint32 planId;
+	uint64 planId;
 	uint64 shardId;
 } PreparedStatementCacheKey;
 
@@ -59,9 +58,9 @@ extern HTAB * PreparedStatementCacheCreate(void);
 extern void PreparedStatementCacheDestroy(HTAB *cache);
 
 /* cache operations */
-extern PreparedStatementCacheEntry * PreparedStatementCacheLookup(
-	HTAB *cache, uint32 planId, uint64 shardId);
-extern PreparedStatementCacheEntry * PreparedStatementCacheInsert(
-	HTAB *cache, uint32 planId, uint64 shardId);
+extern PreparedStatementCacheEntry * PreparedStatementCacheLookup(HTAB *cache, uint64
+																  planId, uint64 shardId);
+extern PreparedStatementCacheEntry * PreparedStatementCacheInsert(HTAB *cache, uint64
+																  planId, uint64 shardId);
 
 #endif /* PREPARED_STATEMENT_CACHE_H */
